+++ src/alloc.c
@@ -4970,7 +4970,7 @@ Garbage collection happens automatically if you cons more than
 				    lcrecord_stats[i].bytes_freed;
 
 			sz = snprintf(buf, sizeof(buf), "%s-storage", name);
-			assert(sz >=0  && sz < sizeof(buf));
+			assert(sz >=0  && (size_t)sz < sizeof(buf));
 			pl = gc_plist_hack(buf, lcrecord_stats[i].bytes_in_use,
 					   pl);
 			/* Okay, simple pluralization check for
@@ -4979,7 +4979,7 @@ Garbage collection happens automatically if you cons more than
                                 sz = snprintf(buf, sizeof(buf), "%ses-freed", name);
 			else
 				sz = snprintf(buf, sizeof(buf), "%ss-freed", name);
-			assert(sz >=0  && sz < sizeof(buf));
+			assert(sz >=0  && (size_t)sz < sizeof(buf));
 			if (lcrecord_stats[i].instances_freed != 0)
 				pl = gc_plist_hack(buf,
 						   lcrecord_stats[i].
@@ -4988,7 +4988,7 @@ Garbage collection happens automatically if you cons more than
 				sz = snprintf(buf, sizeof(buf), "%ses-on-free-list", name);
 			else
 				sz = snprintf(buf, sizeof(buf), "%ss-on-free-list", name);
-			assert(sz >=0  && sz < sizeof(buf));
+			assert(sz >=0  && (size_t)sz < sizeof(buf));
 			if (lcrecord_stats[i].instances_on_free_list != 0)
 				pl = gc_plist_hack(buf,
 						   lcrecord_stats[i].
@@ -4997,7 +4997,7 @@ Garbage collection happens automatically if you cons more than
 				sz = snprintf(buf, sizeof(buf), "%ses-used", name);
 			else
 				sz = snprintf(buf, sizeof(buf), "%ss-used", name);
-			assert(sz >=0  && sz < sizeof(buf));
+			assert(sz >=0  && (size_t)sz < sizeof(buf));
 			pl = gc_plist_hack(buf,
 					   lcrecord_stats[i].instances_in_use,
 					   pl);
